ecb fix

+++ ecb/oldcode/aes_mode_ecb0.py
@@ -150,9 +150,9 @@ def create_modified_image(image_path, output_path):
     img_array = np.array(img)
 
     # Modify one pixel (flip the top-left corner)
-    img_array[0, 0, 0] = (img_array[0, 0, 0] + 1) % 256  # R channel
-    img_array[0, 0, 1] = (img_array[0, 0, 1] + 1) % 256  # G channel
-    img_array[0, 0, 2] = (img_array[0, 0, 2] + 1) % 256  # B channel
+    img_array[0, 0, 0] = np.uint8(int(img_array[0, 0, 0] + 1) % 256)  # R channel
+    img_array[0, 0, 1] = np.uint8(int(img_array[0, 0, 1] + 1) % 256)  # G channel
+    img_array[0, 0, 2] = np.uint8(int(img_array[0, 0, 2] + 1) % 256)  # B channel
 
     # Save the modified image
     modified_img = Image.fromarray(img_array)
